dev: 3s timeoout to check if sst dev is running

+++ platform/functions/bridge/bridge.go
@@ -1,6 +1,8 @@
 package main
 
 import (
+	"bufio"
+	"bytes"
 	"context"
 	"crypto/tls"
 	"encoding/json"
@@ -191,6 +193,15 @@ func run() error {
 	}); token.Wait() && token.Error() != nil {
 		return token.Error()
 	}
+	ack := make(chan struct{})
+	if token := mqttClient.Subscribe(prefix+"/ack", 1, func(c MQTT.Client, m MQTT.Message) {
+		slog.Info("received ack")
+		go func() {
+			ack <- struct{}{}
+		}()
+	}); token.Wait() && token.Error() != nil {
+		return token.Error()
+	}
 
 	if token := mqttClient.Subscribe(prefix+"/kill", 1, func(c MQTT.Client, m MQTT.Message) {
 		slog.Info("received kill message")
@@ -206,11 +217,100 @@ func run() error {
 		<-sigs
 		cancel()
 	}()
-
 	defer func() {
 		mqttClient.Publish(prefix+"/shutdown", 1, false, initPayload).Wait()
 	}()
 
+	for {
+		slog.Info("waiting for next invocation")
+		resp, err := http.Get("http://" + LAMBDA_RUNTIME_API + "/2018-06-01/runtime/invocation/next")
+		if err != nil {
+			return err
+		}
+		body, err := io.ReadAll(resp.Body)
+		if err != nil {
+			return err
+		}
+		requestID := resp.Header.Get("lambda-runtime-aws-request-id")
+		go func() {
+			select {
+			case <-time.After(time.Second * 3):
+				slog.Info("sst dev is not running")
+				cancel()
+				return
+			case <-ack:
+				return
+			}
+		}()
+		slog.Info("got next invocation")
+		for {
+			var conn net.Conn
+			var writer *iot_writer.IoTWriter
+		loop:
+			for {
+				select {
+				case <-ctx.Done():
+					body := map[string]interface{}{
+						"errorMessage": "sst dev is not running",
+						"errorType":    "Error",
+						"stackTrace":   []string{},
+					}
+					data, _ := json.Marshal(body)
+					resp, err := http.Post("http://"+LAMBDA_RUNTIME_API+"/2018-06-01/runtime/invocation/"+requestID+"/error", "application/json", bytes.NewReader(data))
+					if err != nil {
+						return err
+					}
+					slog.Info("error response", "code", resp.StatusCode)
+					return nil
+				case payload := <-requestChan:
+					if len(payload.Data) == 0 {
+						break loop
+					}
+					if writer == nil {
+						writer = iot_writer.New(mqttClient, prefix+"/response/"+payload.ID)
+					}
+					// check if payload.Data is a GET request
+					if payload.Data[0] == 0x47 {
+						resp.Body = io.NopCloser(bytes.NewReader(body))
+						slog.Info("get request")
+						resp.Write(writer)
+						writer.Flush()
+						// drain rest of the request
+						for payload := range requestChan {
+							if len(payload.Data) == 0 {
+								break
+							}
+						}
+						break loop
+					}
+
+					if conn == nil {
+						slog.Info("dialing lambda runtime api")
+						conn, err = net.Dial("tcp", LAMBDA_RUNTIME_API)
+						if err != nil {
+							return err
+						}
+					}
+					_, err = conn.Write(payload.Data)
+					if err != nil {
+						return err
+					}
+				}
+			}
+			if conn == nil {
+				continue
+			}
+			slog.Info("forwarding response")
+			body, err := io.ReadAll(conn)
+			slog.Info("got response", "body", string(body))
+			_, err = io.Copy(writer, bytes.NewReader(body))
+			if err != nil {
+				return err
+			}
+			writer.Flush()
+			break
+		}
+	}
 	for {
 		slog.Info("dialing lambda runtime api")
 		conn, err := net.Dial("tcp", LAMBDA_RUNTIME_API)
@@ -236,15 +336,17 @@ type msg struct {
 }
 
 func forwardRequest(ctx context.Context, requestChan chan iot_writer.ReadMsg, conn net.Conn) (string, error) {
-	count := 0
+	var buffer bytes.Buffer
+	multiWriter := io.MultiWriter(conn, &buffer)
 	for {
 		select {
 		case payload := <-requestChan:
 			if len(payload.Data) == 0 {
-				return payload.RequestID, nil
+				req, _ := http.ReadRequest(bufio.NewReader(&buffer))
+				slog.Info("request", "method", req.Method, "url", req.URL.String())
+				return payload.ID, nil
 			}
-			count++
-			conn.Write(payload.Data)
+			multiWriter.Write(payload.Data)
 			continue
 		case <-ctx.Done():
 			return "", fmt.Errorf("context cancelled")
